预估时间、websocket相关代码

+++ src/main/java/com/bupt/charger/response/ChargeReqResponse.java
@@ -15,5 +15,5 @@ public class ChargeReqResponse extends Resp {
     @JsonProperty("queue_num")
     private String queue;
     @JsonProperty("request_time")
-    private int requestTime = -1;
+    private long requestTime = -1;
 }
